requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
boto3==1.34.47
botocore==1.34.47
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.9
django-allauth==0.62.0
django-bootstrap-v5==1.0.11
django-filter==23.5
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-storages==1.14.2
django-webpack-loader==3.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
psycopg==3.1.17
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
xlrd==2.0.1
